Reference

Contents

Reference

Contents

Index

TestUtils

CPU version (to be deprecated)

Index

Vegas.AbstractTargetDistribution

Vegas.VegasCPU.AbstractProcessSetup

Vegas.TestUtils.get_test_setup

Vegas.VegasCPU._build_psp

Vegas.VegasCPU._compute

Vegas.VegasCPU.coordinate_boundaries

Vegas.VegasCPU.degree_of_freedom

Vegas._compute

Vegas.degrees_of_freedom

Vegas.AbstractTargetDistribution — Type

AbstractTargetDistribution

Abstract base type for target distributions used in sampling.

Concrete subtypes represent functions defined on an N-dimensional real domain and must implement the following interface:

Required Methods

_compute(dist, x::NTuple{N,<:Real})::Real Evaluate the (possibly unnormalized) target distribution at the point x.

degrees_of_freedom(dist)::Int Return the number of degrees of freedom, i.e. the dimensionality of the domain on which the distribution is defined.

Notes

Implementations are expected to be side-effect free and compatible with GPU execution.

Broadcasting over an AbstractTargetDistribution treats the distribution as a scalar.

See also

_compute, degrees_of_freedom

Vegas._compute — Function

_compute(dist::AbstractTargetDistribution,x::NTuple{N,<:Real})::Real

Interface function: Return value of dist computed at x.

Vegas.degrees_of_freedom — Function

degrees_of_freedom(dist::AbstractTargetDistribution)::Int

Interface function: return the degrees of freedom of dist, e.g. the dimension of the domain.

TestUtils

Vegas.TestUtils.get_test_setup — Function

get_test_setup(backend::KernelAbstractions.Backend)

Interface function: return test setup for given backend.

CPU version (to be deprecated)

Vegas.VegasCPU.AbstractProcessSetup — Type

AbstractProcessSetup{P,M,PSL}

Abstract base type for setups related to scattering processes with

P<:AbstractProcessDefinition,

M<:AbstractModelDefinition,

PSL<:AbstractPhaseSpaceLayout.

Interface function to be implemented:

QEDbase.process(stp::AbstractProcessSetup)::P,

QEDbase.model(stp::AbstractProcessSetup)::M,

QEDbase.phase_space_layout(stp::AbstractProcessSetup)::PSL,

_compute(stp::AbstractProcessSetup, psp::AbstractPhaseSpacePoint)::Real.

Optionally, the following interface functions can be implemented:

_build_psp(stp::AbstractProcessSetup,coords::Tuple)::PSP

degree_of_freedom(stp::AbstractProcessSetup)::Int,

coordinate_boundaries(stp::AbstractProcessSetup)::Tuple,

A leading underscore of an interface function means, there is no input validation necessary, because there is a generic implementation of the same function without the underscore, which does basic compiletime checks.

Vegas.VegasCPU._build_psp — Function

_build_psp(stp::AbstractProcessSetup, coords::Tuple)

Vegas.VegasCPU._compute — Function

_compute(stp::AbstractComputationSetup,input::Any)

Vegas.VegasCPU.coordinate_boundaries — Function

coordinate_boundaries(stp::AbstractProcessSetup)

Vegas.VegasCPU.degree_of_freedom — Function

degree_of_freedom(stp::AbstractProcessSetup)